Fix timer app
